api_presenca/requirements.txt
Flask==2.2.2
werkzeug==2.2.2
redis==4.5.5
mysql-connector-python==8.0.28  
